django-to-do-app/requirements.txt
asgiref==3.7.2
awsebcli==3.20.3
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.11.1
botocore==1.23.54
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.3
comtypes==1.1.11
cryptography==36.0.1
distlib==0.3.3
Django==3.2
django-filter==23.2
djangorestframework==3.14.0
docopt==0.6.2
ec2instanceconnectcli==1.0.2
filelock==3.3.1
Flask==2.1.1
future==0.16.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==0.10.0
Js2Py==0.71
Markdown==3.4.3
MarkupSafe==2.1.1
MouseInfo==0.1.3
numpy==1.22.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
pipwin==0.5.2
platformdirs==2.4.0
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==3.0.8
pyperclip==1.8.2
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==0.1.28
pySmartDL==1.3.4
python-dateutil==2.8.2
pyttsx3==2.90
pytweening==1.0.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.4
termcolor==1.1.0
typing_extensions==4.6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.8
virtualenv==20.9.0
wcwidth==0.1.9
Werkzeug==2.1.1
wikipedia==1.4.0
uWSGI>=2.0.19.1,<2.1
gunicorn==20.1.0